requirements.txt
requests==2.33.1
python-dotenv==1.2.2
beautifulsoup4==4.14.3
lxml==6.1.0
pandas==3.0.2
matplotlib==3.10.8
textblob==0.20.0